GLJS-801 Decoupled camera.test.js (internal-1323)
